requirements.txt
absl-py==2.1.0
cloudpickle==3.0.0
contourpy==1.3.0
cycler==0.12.1
deepdiff==7.0.1
distlib==0.3.8
Farama-Notifications==0.0.4
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
grpcio==1.66.2
gymnasium==0.28.1
h5py==3.12.1
jax-jumpy==1.0.0
Jinja2==3.1.4
kiwisolver==1.4.7
llvmlite==0.43.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
minigrid==2.3.1
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
ordered-set==4.1.0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pettingzoo==1.24.3
pillow==10.4.0
platformdirs==4.3.6
protobuf==5.28.2
pygame==2.6.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
sensai-utils==1.2.1
setuptools==75.1.0
six==1.16.0
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tianshou==1.1.0
torch==2.2.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
virtualenv==20.26.6
Werkzeug==3.0.4
